[FC] Using `int` for addresses instead of `ushort`.

+++ src/Beta.Famicom/Boards/Discrete/NROM.cs
@@ -12,7 +12,7 @@ namespace Beta.Famicom.Boards.Discrete
             this.image = image;
         }
 
-        public void R2A03Read(ushort address, ref byte data)
+        public void R2A03Read(int address, ref byte data)
         {
             if ((address & 0x8000) == 0x8000)
             {
@@ -20,9 +20,9 @@ namespace Beta.Famicom.Boards.Discrete
             }
         }
 
-        public void R2A03Write(ushort address, byte data) { }
+        public void R2A03Write(int address, byte data) { }
 
-        public void R2C02Read(ushort address, ref byte data)
+        public void R2C02Read(int address, ref byte data)
         {
             if ((address & 0x2000) == 0x0000)
             {
@@ -30,7 +30,7 @@ namespace Beta.Famicom.Boards.Discrete
             }
         }
 
-        public void R2C02Write(ushort address, byte data)
+        public void R2C02Write(int address, byte data)
         {
             if ((address & 0x2000) == 0x0000)
             {
@@ -38,7 +38,7 @@ namespace Beta.Famicom.Boards.Discrete
             }
         }
 
-        public bool VRAM(ushort address, out int a10)
+        public bool VRAM(int address, out int a10)
         {
             var x = (address >> 10) & image.h;
             var y = (address >> 11) & image.v;
