+++ src/Beta.Famicom/Boards/Discrete/UxROM.cs
@@ -13,7 +13,7 @@ namespace Beta.Famicom.Boards.Discrete
             this.image = image;
         }
 
-        public void R2A03Read(ushort address, ref byte data)
+        public void R2A03Read(int address, ref byte data)
         {
             if ((address & 0x8000) == 0x8000)
             {
@@ -21,7 +21,7 @@ namespace Beta.Famicom.Boards.Discrete
             }
         }
 
-        public void R2A03Write(ushort address, byte data)
+        public void R2A03Write(int address, byte data)
         {
             if ((address & 0x8000) == 0x8000)
             {
@@ -29,7 +29,7 @@ namespace Beta.Famicom.Boards.Discrete
             }
         }
 
-        private int MapR2A03Address(ushort address)
+        private int MapR2A03Address(int address)
         {
             return (address & 0xc000) == 0x8000
                 ? (address & 0x3fff) | (prg_page << 14)
@@ -37,7 +37,7 @@ namespace Beta.Famicom.Boards.Discrete
                 ;
         }
 
-        public void R2C02Read(ushort address, ref byte data)
+        public void R2C02Read(int address, ref byte data)
         {
             if ((address & 0x2000) == 0x0000)
             {
@@ -45,7 +45,7 @@ namespace Beta.Famicom.Boards.Discrete
             }
         }
 
-        public void R2C02Write(ushort address, byte data)
+        public void R2C02Write(int address, byte data)
         {
             if ((address & 0x2000) == 0x0000)
             {
@@ -53,7 +53,7 @@ namespace Beta.Famicom.Boards.Discrete
             }
         }
 
-        public bool VRAM(ushort address, out int a10)
+        public bool VRAM(int address, out int a10)
         {
             var x = (address >> 10) & image.h;
             var y = (address >> 11) & image.v;
