+++ src/Beta.Famicom/Boards/Konami/VRC2.cs
@@ -17,7 +17,7 @@ namespace Beta.Famicom.Boards.Konami
             this.image = image;
         }
 
-        public void R2A03Read(ushort address, ref byte data)
+        public void R2A03Read(int address, ref byte data)
         {
             if ((address & 0x8000) == 0x8000)
             {
@@ -25,7 +25,7 @@ namespace Beta.Famicom.Boards.Konami
             }
         }
 
-        public void R2A03Write(ushort address, byte data)
+        public void R2A03Write(int address, byte data)
         {
             const int a0_shift = 0;
             const int a1_shift = 1;
@@ -91,7 +91,7 @@ namespace Beta.Famicom.Boards.Konami
             throw new CompilerPleasingException();
         }
 
-        public void R2C02Read(ushort address, ref byte data)
+        public void R2C02Read(int address, ref byte data)
         {
             if ((address & 0x2000) == 0x0000)
             {
@@ -99,7 +99,7 @@ namespace Beta.Famicom.Boards.Konami
             }
         }
 
-        public void R2C02Write(ushort address, byte data)
+        public void R2C02Write(int address, byte data)
         {
             if ((address & 0x2000) == 0x0000)
             {
@@ -126,7 +126,7 @@ namespace Beta.Famicom.Boards.Konami
             throw new CompilerPleasingException();
         }
 
-        public bool VRAM(ushort address, out int a10)
+        public bool VRAM(int address, out int a10)
         {
             var x = (address >> 10) & 1;
             var y = (address >> 11) & 1;
