+++ src/Beta.Famicom/CPU/NOI/NoiRegisters.cs
@@ -14,7 +14,7 @@
             this.noi = state.r2a03.noi;
         }
 
-        public void Write(ushort address, byte data)
+        public void Write(int address, byte data)
         {
             switch (address - 0x400c)
             {
