+++ src/Beta.Famicom/Memory/CartridgeConnector.cs
@@ -11,27 +11,27 @@ namespace Beta.Famicom.Memory
             this.board = board;
         }
 
-        public void R2C02Read(ushort address, ref byte data)
+        public void R2C02Read(int address, ref byte data)
         {
             board.R2C02Read(address, ref data);
         }
 
-        public void R2C02Write(ushort address, byte data)
+        public void R2C02Write(int address, byte data)
         {
             board.R2C02Write(address, data);
         }
 
-        public void R2A03Read(ushort address, ref byte data)
+        public void R2A03Read(int address, ref byte data)
         {
             board.R2A03Read(address, ref data);
         }
 
-        public void R2A03Write(ushort address, byte data)
+        public void R2A03Write(int address, byte data)
         {
             board.R2A03Write(address, data);
         }
 
-        public bool VRAM(ushort address, out int a10)
+        public bool VRAM(int address, out int a10)
         {
             return board.VRAM(address, out a10);
         }
